requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
PyMySQL>=1.1.0
Werkzeug>=3.0.0
cryptography>=3.0.0

